backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
